% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rpc.R
\name{otel_grpc_port}
\alias{otel_grpc_port}
\title{Get the OTLP gRPC port from environment}
\usage{
otel_grpc_port()
}
\value{
An integer port number.
}
\description{
Reads \code{OTEL_EXPORTER_OTLP_GRPC_PORT} env var, defaulting to 4317.
}
